replaced all floats with ints

+++ include/gpu.hpp
@@ -5,6 +5,6 @@
 
 using namespace std;
 
-void runGPU(vector<float> &B, vector<float> &C, vector<int> &rows, vector<int> &columns, int numVertices, vector<int> &updates);
+void runGPU(vector<int> &B, vector<int> &C, vector<int> &rows, vector<int> &columns, int numVertices, vector<int> &updates);
 
 #endif
